requirements.txt
django
django-rest-framework
django-cors-headers
unidecode
shortuuid
drf-yasg